% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_hurricane_data.R
\name{clean_hurricane_track}
\alias{clean_hurricane_track}
\title{clean_hurricane_data}
\usage{
clean_hurricane_track(hurricane_data, hurricane_name, location_date)
}
\arguments{
\item{hurricane_data}{The object containing the hurricane data}

\item{hurricane_name}{The name of the hurricane of interest}

\item{location_date}{The date and time point of interest, provided in the format "YYYY-mm-dd-hh-mm"}
}
\value{
tibble object containing the filtered and cleaned data
}
\description{
clean_hurricane_data cleans the data provided to specific hurricane name and datetime
}
\details{
This function takes the data provided and cleans the data by creating a storm_id column, converting
longitude data, and ensuring dates are in the proper format. The data is then modified to gather wind
speed and distance of coverage before filtering by the provided hurricane name and datetime of interest
}
\examples{
\dontrun{
katrina <- clean_hurricane_track(ext_tracks, "Katrina", "2005-08-29-12-00")
head(katrina)
}
}
\seealso{
\link{validate_datetime}
}
