% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hurricane.R
\name{geom_hurricane}
\alias{geom_hurricane}
\title{geom_hurricane}
\usage{
geom_hurricane(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = TRUE,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{aesthetics mapping}

\item{data}{data to be used}

\item{stat}{stat objected with default "identity"}

\item{position}{position object with default "identity"}

\item{na.rm}{handling of NA values, removal default to FALSE}

\item{show.legend}{option to show or hide the legend with default set to FALSE}

\item{inherit.aes}{option to inherit aesthetics specified}

\item{...}{additional parameters passed to the function call}
}
\value{
a layer containing a \verb{Geom*} object responsible for rendering the pie-style graphic
}
\description{
geom_hurricane plots the wind radius and speed of a selected hurricane and datetime
}
\details{
The graphic is a pie-style plot showing the radius and speed of winds from hurricane data using the
ggplot interface.
}
\examples{
\dontrun{
ggplot() +
  geom_hurricane(aes(x = longitude, y = latitude, r_ne = ne, r_se = se, r_nw = nw, r_sw =sw, 
  speed = wind_speed, fill = wind_speed), group = 1))
}
}
