% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurricane_base_map.R
\name{hurricane_base_map}
\alias{hurricane_base_map}
\title{hurricane_base_map}
\usage{
hurricane_base_map(cleaned_data)
}
\arguments{
\item{cleaned_data}{Object containing the cleaned hurricane data}
}
\value{
ggplot2 object showing map location centered on hurricane center
}
\description{
hurricane_base_map uses cleaned hurricane data to get a base map for plotting hurricane data
}
\details{
This function uses cleaned hurricane data to calculate the total area coverage of the specified hurricane,
add a buffer, and create a base map for the hurricane data to be plotted on to
}
\examples{
\dontrun{
baseMap <- hurricane_base_map(katrina)
baseMap
}
}
\seealso{
\link{clean_hurricane_data}
}
