% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_hurricane_data.R
\name{load_hurricane_data}
\alias{load_hurricane_data}
\title{load_hurricane_data}
\usage{
load_hurricane_data(filename)
}
\arguments{
\item{filename}{The full file path to the txt file containing the hurricane data.}
}
\value{
tibble object containing the data
}
\description{
load_hurricane_data loads hurricane data from a fixed width file
}
\details{
This function establishes column widths and column names, checks for the specified txt file using the file name provided, and loads the data.
If the file does not exist, an error is returned stating that the file does not exist.
}
\examples{
\dontrun{
data <- load_hurricane_data("accident_2013.csv.bz2")
head(data)
}
}
