% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_hurricane_data.R
\name{validate_datetime}
\alias{validate_datetime}
\title{validate_datetime}
\usage{
validate_datetime(text)
}
\arguments{
\item{text}{The string of text to be checked}
}
\value{
TRUE or FALSE if the text matches the desrired format
}
\description{
checks string for proper datetime format
}
\details{
This function checks the provided string for datetime format YYYY-mm-dd-hh-mm
}
\examples{
\dontrun{
validate_datetime("2005-08-29-12-00")
}
}
