% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeomHurricane.R
\docType{data}
\name{GeomHurricane}
\alias{GeomHurricane}
\title{GeomHurricane}
\format{
An object of class \code{GeomHurricane} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomHurricane
}
\arguments{
\item{GeomHurricane}{\verb{Geom*} name created to plot hurricane data#'}
}
\value{
pie-style grpahic showing the wind speed and radius covered by the hurricane
}
\description{
Creates a new \verb{Geom*} to show hurricane strength and coverage
}
\details{
The \verb{Geom*} is created by calculating the center of the hurricane at the datetime provided in the data,
converts the radial distance covered for plotting, calculates the area covered by each wind speed portion, and
creates a grob for each wind speed area for each quadrant around the hurricane's center. The function sets the
deafult color scale to yellow = strongest winds, orange = medium winds, and red = weakest winds.
}
\keyword{datasets}
